requirements.txt
discord.py == 2.3.2
python-dotenv==1.0.1